requirements.txt
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==5.4.1
pywin32==224